api/requirements.txt
# Core FastAPI and Server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Database and Auth
supabase>=1.0.0

# Environment and Image Processing
python-dotenv>=1.0.0
Pillow>=9.5.0

# AI Model Runtime
# tensorflow>=2.14.0 # Temporarily removed due to Python 3.13 incompatibility

# Generative AI
google-generativeai>=0.4.0
requests>=2.31.0
